Add extensions to correct protocol.
