+++ Sources/SmokeDynamoDB/DynamoDBCompositePrimaryKeyTable+consistentReadQuery.swift
@@ -78,63 +78,3 @@ public extension DynamoDBCompositePrimaryKeyTable {
                                 consistentRead: self.consistentRead)
     }
 }
- 
-public extension DynamoDBCompositePrimaryKeyTableV2 {
-#if (os(Linux) && compiler(>=5.5)) || (!os(Linux) && compiler(>=5.5.2)) && canImport(_Concurrency)
-    func query<ReturnedType: PolymorphicOperationReturnType>(forPartitionKey partitionKey: String,
-                                                             sortKeyCondition: AttributeCondition?) async throws
-    -> [ReturnedType] {
-        return try await query(forPartitionKey: partitionKey,
-                               sortKeyCondition: sortKeyCondition,
-                               consistentRead: self.consistentRead)
-    }
-    
-    func query<ReturnedType: PolymorphicOperationReturnType>(forPartitionKey partitionKey: String,
-                                                             sortKeyCondition: AttributeCondition?,
-                                                             limit: Int?,
-                                                             exclusiveStartKey: String?) async throws
-    -> ([ReturnedType], String?) {
-        return try await query(forPartitionKey: partitionKey,
-                               sortKeyCondition: sortKeyCondition,
-                               limit: limit,
-                               exclusiveStartKey: exclusiveStartKey,
-                               consistentRead: self.consistentRead)
-    }
-    
-    func query<ReturnedType: PolymorphicOperationReturnType>(forPartitionKey partitionKey: String,
-                                                             sortKeyCondition: AttributeCondition?,
-                                                             limit: Int?,
-                                                             scanIndexForward: Bool,
-                                                             exclusiveStartKey: String?) async throws
-    -> ([ReturnedType], String?) {
-        return try await query(forPartitionKey: partitionKey,
-                               sortKeyCondition: sortKeyCondition,
-                               limit: limit,
-                               scanIndexForward: scanIndexForward,
-                               exclusiveStartKey: exclusiveStartKey,
-                               consistentRead: self.consistentRead)
-    }
-    
-    func monomorphicQuery<AttributesType, ItemType>(forPartitionKey partitionKey: String,
-                                                    sortKeyCondition: AttributeCondition?) async throws
-    -> [TypedDatabaseItem<AttributesType, ItemType>] {
-        return try await monomorphicQuery(forPartitionKey: partitionKey,
-                                          sortKeyCondition: sortKeyCondition,
-                                          consistentRead: self.consistentRead)
-    }
-    
-    func monomorphicQuery<AttributesType, ItemType>(forPartitionKey partitionKey: String,
-                                                    sortKeyCondition: AttributeCondition?,
-                                                    limit: Int?,
-                                                    scanIndexForward: Bool,
-                                                    exclusiveStartKey: String?) async throws
-    -> ([TypedDatabaseItem<AttributesType, ItemType>], String?) {
-        return try await monomorphicQuery(forPartitionKey: partitionKey,
-                                          sortKeyCondition: sortKeyCondition,
-                                          limit: limit,
-                                          scanIndexForward: scanIndexForward,
-                                          exclusiveStartKey: exclusiveStartKey,
-                                          consistentRead: self.consistentRead)
-    }
-#endif
-}
